Todas las tablas ya estan registradas en entidades y ya estan relacionadas todas

+++ src/main/java/org/example/vitalance/dtos/AlimentoDTO.java
@@ -12,7 +12,7 @@ public class AlimentoDTO {
     private Long id;
     private Integer nombre; // o String si corregimos
     private String descripcion;
-    private BigDecimal carbohidrato;
-    private BigDecimal calorias;
-    private BigDecimal indiceGlucemico;
+    private Double carbohidrato;
+    private Double calorias;
+    private Double indiceGlucemico;
 }
